+++ app/src/main/java/com/js/mystore/repository/ProductRepositoryImpl.kt
@@ -4,11 +4,11 @@ import com.js.mystore.data.dao.StoreDao
 import com.js.mystore.model.Product
 
 class ProductRepositoryImpl(private val storeDao: StoreDao) : ProductRepository {
-    override suspend fun getProduct(productId: Int): Product {
+    override suspend fun getProduct(productId: Long): Product {
         return storeDao.getProduct(productId)
     }
 
-    override suspend fun getProductAll(companyId: Int, status: Boolean): List<Product> {
+    override suspend fun getProductAll(companyId: Long, status: Boolean): List<Product> {
         return storeDao.getProductAll(companyId, status)
     }
 
@@ -23,5 +23,4 @@ class ProductRepositoryImpl(private val storeDao: StoreDao) : ProductRepository 
     override suspend fun deleteProduct(product: Product) {
         return storeDao.updateProduct(product)
     }
-
 }
